requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
djangorestframework==3.9.4
docopt==0.6.2
get==2019.4.13
idna==2.8
jedi==0.14.0
parso==0.5.0
post==2019.4.13
prompt-toolkit==2.0.9
ptpython==2.0.4
public==2019.4.13
pycodestyle==2.5.0
Pygments==2.4.2
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
wcwidth==0.1.7
